Reorganizado o código para que os tutoriais sejam submódulos do modulo tutorial
