+++ src/main/java/net/floodlightcontroller/tutorial/two/IPktinHistoryService.java
@@ -1,4 +1,4 @@
-package net.floodlightcontroller.pktinhistory;
+package net.floodlightcontroller.tutorial.two;
  
 import net.floodlightcontroller.core.module.IFloodlightService;
 import net.floodlightcontroller.core.types.SwitchMessagePair;
