+++ src/main/java/net/floodlightcontroller/tutorial/two/PktInHistory.java
@@ -1,4 +1,4 @@
-package net.floodlightcontroller.pktinhistory;
+package net.floodlightcontroller.tutorial.two;
 import net.floodlightcontroller.core.types.*;
 import net.floodlightcontroller.restserver.IRestApiService;
 import net.floodlightcontroller.util.ConcurrentCircularBuffer;
