+++ src/main/java/net/floodlightcontroller/tutorial/two/PktInHistoryResource.java
@@ -1,4 +1,4 @@
-package net.floodlightcontroller.pktinhistory;
+package net.floodlightcontroller.tutorial.two;
  
 import java.util.ArrayList;
 import java.util.List;
