+++ src/main/java/net/floodlightcontroller/tutorial/two/PktInHistoryWebRoutable.java
@@ -1,4 +1,4 @@
-package net.floodlightcontroller.pktinhistory;
+package net.floodlightcontroller.tutorial.two;
  
 import org.restlet.Context;
 import org.restlet.Restlet;
