+++ src/main/java/net/floodlightcontroller/util/UtilLog.java
@@ -4,7 +4,7 @@ import org.fusesource.jansi.Ansi;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import net.floodlightcontroller.deva.DevaProcessPacketIn;
+import net.floodlightcontroller.tutorial.one.DevaProcessPacketIn;
 
 public class UtilLog {
