+++ src/main/resources/floodlightdefault.properties
@@ -20,9 +20,9 @@ net.floodlightcontroller.devicemanager.internal.DeviceManagerImpl,\
 net.floodlightcontroller.accesscontrollist.ACL,\
 net.floodlightcontroller.statistics.StatisticsCollector,\
 net.floodlightcontroller.hasupport.HAController,\
-net.floodlightcontroller.deva.MACTracker,\
-net.floodlightcontroller.pktinhistory.PktInHistory,\
-net.floodlightcontroller.deva.DevaProcessPacketIn
+net.floodlightcontroller.tutorial.one.MACTracker,\
+net.floodlightcontroller.tutorial.two.PktInHistory,\
+net.floodlightcontroller.tutorial.one.DevaProcessPacketIn
 org.sdnplatform.sync.internal.SyncManager.authScheme=CHALLENGE_RESPONSE
 org.sdnplatform.sync.internal.SyncManager.keyStorePath=/etc/floodlight/myKey.jceks
 org.sdnplatform.sync.internal.SyncManager.dbPath=/var/lib/floodlight/
